+++ apps/sim/lib/core/utils/urls.ts
@@ -170,6 +170,21 @@ export function getBrowserOrigin(): string | null {
 }
 
 /**
+ * Validates that a URL uses an http(s) scheme before it is opened in a new window.
+ * Rejects `javascript:`, `data:`, `blob:`, `vbscript:`, and other schemes that could
+ * execute script in the chat origin, since `file.url` originates from untrusted
+ * workflow/agent output.
+ */
+export function isSafeHttpUrl(url: string): boolean {
+  try {
+    const parsed = new URL(url, getBrowserOrigin() ?? undefined)
+    return parsed.protocol === 'http:' || parsed.protocol === 'https:'
+  } catch {
+    return false
+  }
+}
+
+/**
  * Returns the socket server URL for server-side internal API calls.
  * Reads from SOCKET_SERVER_URL with a localhost fallback for development.
  */
